requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
python-dotenv
pydantic

langchain-core
langchain-ollama
langchain-openai
langchain-anthropic
langchain-google-genai

duckduckgo-search

opencv-python
ultralytics

sounddevice
numpy
torch
faster-whisper
TTS